Start e2e `collect_images.py` helper file
This file will parse `targets.yml` for the list of images with targets to test.

The testing directory is found with the TF_TESTING_IMG environment variable.
An alternative way to load this is done by the python-dotenv package.
Putting `TF_TESTING_IMG="/some/path"` in a `.env` file in the root directory of the repo will enable this.
This will not override the environment variable however if it is already set.