Added image property to Target
Also added Pillow (PIL) as a dependency since (at least for now) the
library is expected to return a PIL.Image object back to the user so
they can do whatever they want with it.

Added image_width and image_height properties as well, these are
intended to be updated as the process goes on. At the end, we'll take
this data and send back a cropped version of the source image (if
possible).