Ensure that the Target and Blob strings can be eval-ed
The image had to be set to None for Blob since it's a required field,
and that images are not being stored in the string representation. A
line for each test case was added that asserts that eval-ing the `repr`
of the string is equivalent by testing the `repr` again.

Made some code style changes to fit PEP8 since it seems that the style
checker is not covering the target_finder/ directory