Add code style checks to the target_finder/ directory
Originally these were only being applied to the test folder, which isn't
exactly very useful. Instead of using a plugin attached to pep8, this
just runs right before. That is, if the code style checks fail, unit
tests will not run. Note that the pep8 package was renamed to
pycodestyle.